feat(logs): Put logs page behind feature flag
For users that accidentally land on this page from links like the blog.
