+++ static/app/routes.tsx
@@ -1812,7 +1812,7 @@ function buildRoutes() {
 
   const logsChildRoutes = (
     <Fragment>
-      <IndexRoute component={make(() => import('sentry/views/explore/logs'))} />
+      <IndexRoute component={make(() => import('sentry/views/explore/logs/content'))} />
       {traceViewRoute}
     </Fragment>
   );
@@ -1885,7 +1885,9 @@ function buildRoutes() {
       >
         {releasesChildRoutes}
       </Route>
-      <Route path="logs/">{logsChildRoutes}</Route>
+      <Route path="logs/" component={make(() => import('sentry/views/explore/logs'))}>
+        {logsChildRoutes}
+      </Route>
       <Route
         path="saved-queries/"
         component={make(() => import('sentry/views/explore/savedQueries'))}
